v1.0.4 - Pregenerated download tools, ground truth updates, render throughput improvements

+++ process_mesh/utils.hpp
@@ -37,6 +37,6 @@ Eigen::Tensor<unsigned char, 3> compute_flow_viz(const Eigen::Tensor<double, 3> 
 
 Eigen::Tensor<unsigned char, 3> to_color_map(const Eigen::Tensor<double, 2> &input_image, const double &min_percentile, const double &max_percentile, const double &minval=1e-3, const clrmap &type=clrmap::Jet);
 
-Eigen::Tensor<unsigned char, 3> to_color_map(const Eigen::Tensor<int, 2> &input_image);
+Eigen::Tensor<unsigned char, 3> to_color_map(const Eigen::Tensor<long, 2> &input_image);
 
 const std::vector<loop_obj> image_iterator(const int width, const int height, const std::string desc="");
